--- src/pages/Authorization/signIn.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8"/>
-    <meta
-            name="viewport"
-            content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge"/>
-    <title>Sign In</title>
-</head>
-<body>
-<div class="container">
-    <div class="header">
-        <h2>Sign in</h2>
-    </div>
-    <form id="form" class="form">
-        <div class="form-control">
-            <label for="username">Username</label>
-            <input type="text" placeholder="Enter your username" id="username"/>
-            <i class="check-circle"></i>
-            <i class="exclamation-circle"></i>
-            <small id="ErrorLogin">Error message</small>
-        </div>
-
-        <div class="form-control name">
-            <label for="password">Password</label>
-            <input
-                    type="password"
-                    placeholder="Enter your password"
-                    id="password"
-            />
-        </div>
-        <a href="../dist/signUp.html" class="sign-redirect">Don`t have account?</a>
-        <button>Submit</button>
-    </form>
-</div>
-</body>
-</html>
