--- src/pages/Registration/signUp.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8"/>
-    <meta
-            name="viewport"
-            content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
-    <meta http-equiv="X-UA-Compatible" content="ie=edge"/>
-    <title>Document</title>
-</head>
-<body>
-<div class="container">
-    <div class="header">
-        <h2>Create Account</h2>
-    </div>
-    <form id="form" class="form">
-        <div class="form-control">
-            <label for="username">Username</label>
-            <input type="text" placeholder="Enter your username" id="username"/>
-            <i class="check-circle"></i>
-            <i class="exclamation-circle"></i>
-            <small>Error message</small>
-        </div>
-
-        <div class="form-control">
-            <label for="password">Password</label>
-            <input
-                    type="password"
-                    placeholder="Enter your password"
-                    id="password"
-            />
-            <i class="check-circle"></i>
-            <i class="exclamation-circle"></i>
-            <small>Error message</small>
-        </div>
-
-        <div class="form-control">
-            <label for="first_name">First Name</label>
-            <input
-                    type="text"
-                    placeholder="Enter your first name"
-                    id="first_name"
-            />
-            <i class="check-circle"></i>
-            <i class="exclamation-circle"></i>
-            <small>Error message</small>
-        </div>
-
-        <div class="form-control">
-            <label for="last_name">Last name</label>
-            <input
-                    type="text"
-                    placeholder="Enter your last name"
-                    id="last_name"
-            />
-            <i class="check-circle"></i>
-            <i class="exclamation-circle"></i>
-            <small id="sendError">Error message</small>
-        </div>
-        <a href="./dist/signIn.html" class="sign-redirect">Already have account?</a>
-        <button>Submit</button>
-    </form>
-</div>
-</body>
-</html>
