+++ src/pages/mainPage/mainScreen.html
@@ -10,14 +10,14 @@
 
 <body>
 <header>
-    <div class="logo"><a  class="mainA" href=""><img src="./img/logo.svg" alt="devEducation" class="imgLogo"></a></div>
+    <div class="logo"><a class="mainA" href=""><img src="./src/img/logo.jpg" alt="devEducation" class="imgLogo"></a></div>
     <section class="donate"><a class="mainA" href="https://github.com/richerstars/Team_Project2">Donate</a>
     </section>
     <div class="containerBetween"></div>
     <nav class="headerNav">
-        <div class="search"><img src="./img/logoSearch.svg" alt="logoSearch"></div>
-        <div class="filters"><input type="image" src="./img/logoFilters.svg" alt="logoFilters" id="filters"></div>
-        <div class="logoUsers"><input type="image" src="./img/logoUser.svg" alt="logoUser" id="auth"></div>
+        <div class="search"><img src="./src/img/logoSearch.svg" alt="logoSearch"></div>
+        <div class="filters"><input type="image" src="./src/img/logoFilters.svg" alt="logoFilters" id="filters"></div>
+        <div class="logoUsers"><input type="image" src="./src/img/logoUser.svg" alt="logoUser" id="auth"></div>
     </nav>
 </header>
 <main class="wrapper">
@@ -113,9 +113,9 @@
     <section class="filmsShowMore">
     </section>
     <template id="filmsShowMore">
-<!--        <div class="containerHeaderFilms">-->
-<!--            <h3 class="headerFilm"></h3>-->
-<!--        </div>-->
+        <!--        <div class="containerHeaderFilms">-->
+        <!--            <h3 class="headerFilm"></h3>-->
+        <!--        </div>-->
         <div class="containerFilmsItem">
             <div class="filmsItem"><img class="imgFilmsItem" src="./img/logo.jpg">
                 <p class="descriptionFilm"></p>
@@ -165,7 +165,7 @@
             </div>
 
             <div class="form-control">
-                <label  class="mainLabel" for="last_name">Last name</label>
+                <label class="mainLabel" for="last_name">Last name</label>
                 <input
                         type="text"
                         placeholder="Enter your last name"
@@ -176,7 +176,6 @@
                 <small id="sendError">Error message</small>
             </div>
             <button class="SignUpBtn" id="checkSignIn">Already have account?</button>
-<!--            <a href="./dist/signIn.html" class="sign-redirect" class="mainA">Already have account?</a>-->
             <button class="SignUpBtn" id="setSignUpBtn">Submit</button>
         </form>
     </div>
@@ -202,7 +201,6 @@
                 />
             </div>
             <button class="SignUpBtn" id="checkSignUp">Don`t have account?</button>
-<!--            <a href="../dist/signUp.html" class="sign-redirect">Don`t have account?</a>-->
             <button id="setSignInBtn">Submit</button>
         </form>
     </div>
