+++ webpack.config.js
@@ -1,13 +1,11 @@
 const path = require('path')
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-// const CopyPlugin = require("copy-webpack-plugin");
+ // const CopyWebpackPlugin = require("copy-webpack-plugin");
 
 module.exports = {
     entry: {
         main: './src/pages/mainPage/index.js',
-        // signUp: './src/pages/Registration/signUp.js',
-        // signIn: './src/pages/Authorization/signIn.js'
     },
     output: {
         path: path.resolve(__dirname, './dist'),
@@ -18,7 +16,7 @@ module.exports = {
     mode: 'development',
     module: {
 
-        rules: [{test: /\.css$/, use: [MiniCssExtractPlugin.loader, {loader: "css-loader", options: {url: true}},],},
+        rules: [{test: /\.css$/i, use: [MiniCssExtractPlugin.loader, {loader: "css-loader", options: {url: true}},],},
             {
                 test: /\.(png|jpe?g|gif|svg|jpg)$/i,
                 use: [{loader: 'file-loader', options: {name: './img/[name].[ext]',},},],
@@ -47,17 +45,7 @@ module.exports = {
             template: "./src/pages/mainPage/mainScreen.html",
             chunks: ['main']
         }),
-        // new HtmlWebpackPlugin({
-        //     filename: "signIn.html",
-        //     template: './src/pages/Authorization/signIn.html',
-        //     chunks: ['signIn']
-        // }),
-        // new HtmlWebpackPlugin({
-        //     filename: "index.html",
-        //     template: './src/pages/Registration/signUp.html',
-        //     chunks: ['signUp']
-        // }),
-        // new CopyPlugin({
+        // new CopyWebpackPlugin({
         //     patterns: [
         //         {
         //             from: path.resolve(__dirname, './src/img'),
